infra-ia/requirements.txt
# BerinIA - Liste des dépendances

# Base
python-dotenv==1.0.0
pydantic==2.5.2
colorama==0.4.6

# API OpenAI
openai==1.3.7
httpx==0.25.2

# Gestion des tâches
schedule==1.2.1
apscheduler==3.10.4

# Outils de traitement
tqdm==4.66.1
orjson==3.9.10
numpy==1.26.2
python-dateutil==2.8.2

# Serveur web pour les webhooks
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Mémoire vectorielle
qdrant-client==1.6.4
sentence-transformers==2.2.2

# Email et SMS
requests==2.31.0
twilio==8.10.0

# Logging
structlog==23.2.0
rich==13.7.0

# Outils de développement
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
